requirements.txt
openai
python-dotenv
pyqrcode
Pillow